also updated selenium_basics_1 HW
